pyproject.toml
[tool.poetry]
name = "djangostarter"
version = "0.1.0"
description = "Django Starter Template"
authors = ["NumanIbnMazid <numanibnmazid@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.1"
django-environ = "^0.8.1"
psycopg2-binary = "^2.9.3"
whitenoise = "^5.3.0"
django-debug-toolbar = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
